@@ -11,9 +11,9 @@ const store = new Vuex.Store({
   modules: {
     counter,
   },
-  strict: debug,
+  strict: (TNS_ENV === 'debug'),
 });
 
 Vue.prototype.$store = store;
 
-module.exports = store;
+export default store;
