@@ -1,26 +1,28 @@
-const validateAppId = (app_id) => {
-  let parts = app_id.split('.');
+const validateAppId = app_id => {
+  let parts = app_id.split('.')
   if (parts.length < 2) {
-    return 'App ID must contain two or more strings, separated by a dot.';
+    return 'App ID must contain two or more strings, separated by a dot.'
   }
-  if (parts.some(part => {
-      return !/^[a-zA-Z]\w+$/i.test(part);
-    })) {
-    return 'Each string must start with a letter and should contain only letters and numbers.';
+  if (
+    parts.some(part => {
+      return !/^[a-zA-Z]\w+$/i.test(part)
+    })
+  ) {
+    return 'Each string must start with a letter and should contain only letters and numbers.'
   }
   if (!/^[a-z]/.test(app_id[0])) {
-    return 'App ID must start with a lowercase letter.';
+    return 'App ID must start with a lowercase letter.'
   }
 
-  return true;
-};
+  return true
+}
 
-const validateVersion = (version) => {
+const validateVersion = version => {
   if (!/^(\d+\.)(\d+\.)(\d)$/.test(version)) {
-    return 'Version must have major, minor and patch numbers.';
+    return 'Version must have major, minor and patch numbers.'
   }
-  return true;
-};
+  return true
+}
 
 module.exports = {
   prompts: {
@@ -95,15 +97,15 @@ module.exports = {
     },
   },
   helpers: {
-    androidVersionCode: (version) => {
-      const parts = version.split('.');
-      return parts[0] + '0' + parts[1] + '0' + parts[2];
+    androidVersionCode: version => {
+      const parts = version.split('.')
+      return parts[0] + '0' + parts[1] + '0' + parts[2]
     },
   },
   filters: {
-    'src/router/**/*': 'router',
-    'src/components/Home.vue': 'router',
-    'src/store/**/*': 'store',
-    'src/components/Counter.vue': 'store',
+    'app/router/**/*': 'router',
+    'app/components/Home.vue': 'router',
+    'app/store/**/*': 'store',
+    'app/components/Counter.vue': 'store',
   },
-};
+}
