@@ -9,8 +9,8 @@
     <StackLayout class="hello-world">
       <Label class="body" textWrap=true text="This is a hello world component, tap the button if you dare"/>
 
-      <Button class="btn btn-primary" @tap="surprise = !surprise" text="Tap me!"/>
-      <Image v-if="surprise" src="~/images/NativeScript-Vue.png"/>
+      <Button class="btn btn-primary" @tap="surprise = !surprise" :text="title"/>
+      <Image v-if="surprise" src="~/assets/images/NativeScript-Vue.png"/>
 
     </StackLayout>
 
@@ -22,6 +22,7 @@
     data () {
       return {
         surprise: false,
+        title: "Tap me!"
       };
     },
   };
